Solución de tipo release integer

+++ backend/src/models/game_model.py
@@ -10,7 +10,7 @@ class GameModel(db.Model):
 	platform = db.Column(db.String(20), nullable=False, index=True)
 	gender = db.Column(db.String(30), nullable=False)
 	pegi = db.Column(db.String(3), nullable=False)
-	release = db.Column(db.Integer(4), nullable=False)
+	release = db.Column(db.Integer(), nullable=False)
 	price = db.Column(db.Numeric(precision=10, scale=2, asdecimal=True), nullable=False)
 	img_url = db.Column(db.String(150), unique=True, nullable=False)
 	stock = db.Column(db.Integer, nullable=False, default=20)
